Chapter I: The Dawn of the Jurassic Idea
Hello whoever may be reading this. This is my story based on an alternate universe i had created inspired by my childhood imagination following a character I created in my dreams. This follows his journey alongside John Hammond as they create Jurassic Park. These chapters set up the foundation and start off from the dream and end at the unfortunate accident of April 12th where a gatekeeper is killed during a transfer leading to serious meeting for accountability and the acquisition of a paleontologist and mathematician alongside an archaeologist.